--- .env
@@ -1,6 +1,0 @@
-SECRET=hawks
-
-dbUser=thunderhawk
-password=utAmmxqSpZWcTqPj
-DATABASE_URI=mongodb+srv://thunderhawk:utAmmxqSpZWcTqPj@cluster0.ww8vv.mongodb.net/kind?retryWrites=true&w=majority
-
