+++ models/User.js
@@ -21,10 +21,6 @@ const userSchema = new Schema({
     type: String,
     required: true
   },
-  receiveDailyMessages: {
-    type: Boolean,
-    default: false
-  },
   messages: [{
     type: mongoose.Schema.Types.ObjectId,
     ref: 'Message'
